Added custom user-agent support to mitigate CloudFlare blocking
Occasionally, CloudFlare may block the script from accessing and
validating the challenge due to the default user-agent string looking
suspicious to CloudFlare, resulting in a HTTP 403 Forbidden.
(The default user-agent usually looks like "Python-urllib/x.x", where
x.x is the Python version.) This is mitigated by allowing the user to
set a browser user-agent, which in turn makes the request succeed.