Adds possibility to pass csr as string instead of as file
To achieve this, code is refactored:
The new function get_crt_from_csr uses the existing code, but expects
the csr as a string instead of as a file name.
The new function get_crt_from_file provides the old functionality,
but the csr file is only read once from disk.