Added WellKnownCheck=True keyword parameter to main function.
In some cases when running behind SNAT or on a LAN the IP address that is pointed to by the common name is not the same as the IP address that it is served on locally.  Because of this we need a way to disable this WellKnownCheck without affecting the default behavior.