Improve processing.
Improve the process of obtaining nonce to save the number of requests that needed to be sent to the server. Improve the process of handling the authorization requests. Improve the process of polling the challenges. Modify the code for obtaining the URLs from the directory. Added a static header for all outgoing requests.